syba & bb into synthesizability.py, correction of nb texts & qed cutoff
