+++ kinfraglib/filters/__init__.py
@@ -1,11 +1,10 @@
 from . import (
-    syba,
     plots,
     check,
     prefilters,
-    building_blocks,
     analysis,
     utils,
     unwanted_substructures,
     drug_likeness,
+    synthesizability
 )
